MINOR: mux-h1: Simplify handling of 1xx responses
Because of previous changes on http tunneling, the synchronization of the
transaction can be simplified. Only the check on intermediate messages remains
and it only concerns the response path.

This patch must be backported to 1.9. It is not strictly speaking required but
it will ease futur backports.

(cherry picked from commit cdc90e9175b476c15d811c0627b1471f5b3177c9)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>