BUG/MEDIUM: spoe: Queue message only if no SPOE applet is attached to the stream
If a SPOE applet is already attached to a stream to handle its messages, we must
not queue them. Otherwise it could be handled by another applet leading to
errors. This happens with fragmented messages only. When the first framgnent is
sent, the SPOE applet sending it is attached to the stream. It should be used to
send all other fragments.

This patch must be backported to 1.9 and 1.8.

(cherry picked from commit 3e86cec05ec9cf848abd8f9a79928410874b778d)
[wla: s/_HA_ATOMIC_ADD/HA_ATOMIC_ADD/ in context]
Signed-off-by: William Lallemand <wlallemand@haproxy.org>