BUG/MEDIUM: stream_interface: Don't bother doing chk_rcv/snd if not connected.
If the interface is not in state SI_ST_CON or SI_ST_EST, don't bother
trying to send/recv data, we can't do it anyway, and if we're in SI_ST_TAR,
that may lead to adding the SI_FL_ERR flag back on the stream_interface,
while we don't want it.

This should be backported to 1.9.

(cherry picked from commit b2fc04ebef4e7eda23ff445a8b2622f2bb75dec1)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>