BUG/MINOR: mux-h1: Process input even if the input buffer is empty
It is required, at least, to add the EOM block and finish the message when the
TCP splicing was used to send all data. Otherwise, there is no way to finish the
parsing.

This patch must be backported to 1.9.

(cherry picked from commit 91f77d59997665c3e6de4b4dc6ca3cdba4e25455)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>