MINOR: mux-h1: Handle read0 during TCP splicing
It avoids a roundtrip with underlying I/O callbacks to do so. If a read0 is
handled at the end of h1_rcv_pipe(), the flag CS_FL_REOS is set on the
conn_stream. And if there is no data in the pipe, the flag CS_FL_EOS is also
set.

This path may be backported to 1.9.

(cherry picked from commit 038ad8123bf51c1b3e4da0bba4be75e33714a5fc)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>