MINOR: peers: adds counters on show peers about tasks calls.
This patch adds a counter of calls on the orchestator peers task
and a counter on the tasks linked to applet i/o handler for
each peer.

Those two counters are useful to detect if a peer sync is active
or frozen.

This patch is related to the commit:
  "MINOR: peers: Add a new command to the CLI for peers."
and should be backported with it.

(cherry picked from commit 0bbec0fa340432944d64575e08fc41d034b5ffc9)
Signed-off-by: Willy Tarreau <w@1wt.eu>